+++ src/finlang/Transpiler/C99ClsEnumInterface.cs
@@ -112,7 +112,7 @@ public class C99ClsEnumInterface
     internal void AddHeaderFqnDependency(ITypeSymbol type)
     {
         // if the type is a c_array, we need to add the dependency of the type it contains
-        if (type is INamedTypeSymbol namedType && namedType.Name == "c_array")
+        if (type is INamedTypeSymbol namedType && namedType.IsCArray())
         {
             type = namedType.TypeArguments.Single();
         }